+++ src/location.c
@@ -4,7 +4,7 @@ size_t location_sizeof(void) {
   return sizeof(location_T);
 }
 
-location_T* location_init(int line, int column) {
+location_T* location_init(size_t line, size_t column) {
   location_T* location = (location_T*) malloc(location_sizeof());
   location->line = line;
   location->column = column;
@@ -12,11 +12,11 @@ location_T* location_init(int line, int column) {
   return location;
 }
 
-int location_line(location_T* location) {
+size_t location_line(location_T* location) {
   return location->line;
 }
 
-int location_column(location_T* location) {
+size_t location_column(location_T* location) {
   return location->column;
 }
